# onebox_technical_test-openapi.yaml
openapi: "3.1.0"
info:
  title: "onebox_technical_test API"
  description: "onebox_technical_test API"
  version: "1.0.0"
servers:
  - url: "http://localhost:8080"
tags:
  - name: "Cart Controller"
    description: "Operations related to shopping carts, including creation, viewing,\
    \ modifying cart contents, and deletion."
paths:
  /api/v1/carts/{cartId}:
    put:
      summary: "Adds a product to a cart"
      description: "This endpoint adds a product to an existing cart. You must provide\
        \ the cart ID and a ProductDto in the body."
      operationId: ""
      parameters:
        - name: "cartId"
          in: "path"
          required: true
          description: "The ID of the cart to add the product to"
          schema:
            type: "string"
            format: "uuid"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductDto"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Cart"
  /api/v1/carts/{id}:
    get:
      summary: "Gets a cart by its ID"
      description: "This endpoint retrieves the cart with the given ID. You must provide\
        \ a valid cart ID to fetch the details."
      operationId: ""
      parameters:
        - name: "id"
          in: "path"
          required: true
          description: "ID of the cart to retrieve"
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Cart"

    delete:
      summary: "Deletes a cart by its ID"
      description: "This endpoint deletes a cart. You must provide the cart ID."
      operationId: ""
      parameters:
        - name: "cartId"
          in: "path"
          required: true
          description: "The ID of the cart to delete"
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DeleteResponse"

  /api/v1/carts:
    post:
      summary: "Creates a new cart"
      description: "This endpoint creates a new, empty cart."
      operationId: ""
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Cart"

components:
  schemas:
    ProductDto:
      type: "object"
      properties:
        description:
          type: "string"
        amount:
          type: "integer"
          format: "int32"
    Product:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        description:
          type: "string"
        amount:
          type: "integer"
          format: "int32"
    Instant:
      type: "object"
      properties: { }
    Cart:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        products:
          type: "array"
          items:
            $ref: "#/components/schemas/Product"
        lastUpdate:
          $ref: "#/components/schemas/Instant"
    DeleteResponse:
      type: "object"
      properties:
        status:
          type: "string"
        message:
          type: "string"